Split code into files
